Moved iomediator_constants into iomediator package. Renamed to constants, because the iomediator prefix is given by the package name.
